Release shared calendars and nested group compatibility
Shared Calendars
Each target group now creates a shareable calendar that replicates every OOO
event with the creator's name and event title. This is in addition to the
calendar that creates a single "all-day" event aggregating all events for the
day, which is used only to generate the email digest. Perhaps some day one
calendar will be able to serve both purposes.

Nested Groups
Users will now be included in the target group even if they are indirect/nested
members of the parent group. Google APIs only consider direct members of a
group to be "members", while many organizations use sub-groups to manage
membership of large aggregate groups, such as "team@" or "product@" or "tech@"
etc.

To solve this, the script creates two arrays: a flattened tree of all groups
eventually contained by the target group, and a list of all groups a user
belongs to. If any of the user's memberships are found in the target group
array, the match is true, and the user is considered part of the group.

The flattened tree of all target groups is meant to be triggered separately
for maintenance, e.g., daily, at a different time.